+++ src/jvmMain/kotlin/ui/stateview/DisplayList.kt
@@ -1,0 +1,61 @@
+package ui.stateview
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.VerticalScrollbar
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.defaultScrollbarStyle
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.rememberScrollbarAdapter
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+
+/**
+ * Display a list in a column with indexes.
+ * With a scrollbar on the right side.
+ */
+@Composable
+fun <T> DisplayList(list: List<T>) {
+    val state = rememberLazyListState()
+
+    Box(
+        modifier = Modifier.fillMaxSize(),
+    ) {
+        LazyColumn(
+            Modifier.fillMaxSize().border(
+                border = BorderStroke(1.dp, MaterialTheme.colorScheme.outline),
+                shape = MaterialTheme.shapes.medium,
+            ).clip(MaterialTheme.shapes.medium),
+            state = state,
+        ) {
+            itemsIndexed(list) { index, it ->
+                val color =
+                    if (index % 2 == 0) MaterialTheme.colorScheme.surfaceVariant else MaterialTheme.colorScheme.surface
+                Text("$index: $it", Modifier.fillMaxWidth().background(color).padding(8.dp))
+            }
+        }
+
+        VerticalScrollbar(
+            modifier = Modifier.align(Alignment.CenterEnd).fillMaxHeight().padding(all = 4.dp),
+            adapter = rememberScrollbarAdapter(
+                scrollState = state,
+            ),
+            style = defaultScrollbarStyle().copy(
+                unhoverColor = MaterialTheme.colorScheme.outline,
+                hoverColor = MaterialTheme.colorScheme.onSurface,
+            ),
+        )
+    }
+}
