+++ src/jvmMain/kotlin/viewmodel/DebuggerViewModel.kt
@@ -58,15 +58,9 @@ class DebuggerViewModel private constructor(val file: File, stateTracker: StateT
     private var currentEvaluation: Int = 0
 
     init {
-        val blockEvaluations = codeAttributes[currentCodeAttribute].blockEvaluations
-        val blockEvaluation = blockEvaluations[currentBlockEvaluation]
-        currentExceptionHandler = blockEvaluation.exceptionHandlerInfo
-        evaluation = blockEvaluation.evaluations[currentEvaluation]
-        currentBlockEvaluationStack = blockEvaluation.branchEvaluationStack
-        hasNext = currentCodeAttribute < codeAttributes.size - 1 ||
-            currentBlockEvaluation < blockEvaluations.size - 1 ||
-            currentEvaluation < blockEvaluation.evaluations.size - 1
-        hasPrevious = currentCodeAttribute > 0 || currentBlockEvaluation > 0 || currentEvaluation > 0
+        update()
+        hasNext = hasNext()
+        hasPrevious = hasPrevious()
     }
 
     private fun update() {
@@ -75,12 +69,6 @@ class DebuggerViewModel private constructor(val file: File, stateTracker: StateT
         currentExceptionHandler = blockEvaluation.exceptionHandlerInfo
         evaluation = blockEvaluation.evaluations[currentEvaluation]
         currentBlockEvaluationStack = blockEvaluation.branchEvaluationStack
-
-        hasNext = currentCodeAttribute < codeAttributes.size - 1 ||
-            currentBlockEvaluation < blockEvaluations.size - 1 ||
-            currentEvaluation < blockEvaluation.evaluations.size - 1
-
-        hasPrevious = currentCodeAttribute > 0 || currentBlockEvaluation > 0 || currentEvaluation > 0
     }
 
     private fun previousInstruction() {
@@ -135,6 +123,8 @@ class DebuggerViewModel private constructor(val file: File, stateTracker: StateT
             Display.EVALUATIONS -> nextEvaluation()
             Display.RESULTS -> nextInstruction()
         }
+        hasNext = hasNext()
+        hasPrevious = hasPrevious()
     }
 
     fun previous() {
@@ -142,6 +132,38 @@ class DebuggerViewModel private constructor(val file: File, stateTracker: StateT
             Display.EVALUATIONS -> previousEvaluation()
             Display.RESULTS -> previousInstruction()
         }
+        hasNext = hasNext()
+        hasPrevious = hasPrevious()
+    }
+
+    private fun hasNext(): Boolean {
+        return when (display) {
+            Display.EVALUATIONS -> {
+                val blockEvaluations = codeAttributes[currentCodeAttribute].blockEvaluations
+                val blockEvaluation = blockEvaluations[currentBlockEvaluation]
+                currentCodeAttribute < codeAttributes.size - 1 ||
+                    currentBlockEvaluation < blockEvaluations.size - 1 ||
+                    currentEvaluation < blockEvaluation.evaluations.size - 1
+            }
+
+            Display.RESULTS -> currentInstruction < codeAttributes[currentCodeAttribute].instructions.size - 1
+        }
+    }
+
+    private fun hasPrevious(): Boolean {
+        return when (display) {
+            Display.EVALUATIONS -> currentCodeAttribute > 0 || currentBlockEvaluation > 0 || currentEvaluation > 0
+            Display.RESULTS -> currentInstruction > 0
+        }
+    }
+
+    fun switchDisplay() {
+        display = when (display) {
+            Display.EVALUATIONS -> Display.RESULTS
+            Display.RESULTS -> Display.EVALUATIONS
+        }
+        hasNext = hasNext()
+        hasPrevious = hasPrevious()
     }
 
     /**
